requirements.txt
langchain_google_genai
langchain
langchain_core
langgraph
python-dotenv
requests
fastapi
langfuse
uvicorn[standard]